fix(transactions): resolve currency formatting issues

+++ database/migrations/2026_04_24_133235_create_accounts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->string('name'); // BCA, Cash, E-Wallet
-            $table->decimal('balance', 15, 2)->default(0);
+            $table->unsignedBigInteger('balance')->default(0);
             $table->timestamps();
         });
     }
